Catch using embedded objects on sets in annotation processor
